fix: fix the vchart-extension packaged artifacts contained an extra vrender
