+++ docs/package.json
@@ -19,8 +19,8 @@
     "@visactor/vchart-theme": "~1.6.6",
     "@visactor/vmind": "1.2.4-alpha.5",
     "@visactor/vutils": "~1.0.12",
-    "@visactor/vrender": "~1.0.21",
-    "@visactor/vrender-kits": "~1.0.21",
+    "@visactor/vrender": "~1.0.23",
+    "@visactor/vrender-kits": "~1.0.23",
     "@visactor/vtable": "1.19.0-alpha.0",
     "@visactor/vtable-editors": "1.19.0-alpha.0",
     "@visactor/vtable-gantt": "1.19.0-alpha.0",
