+++ packages/vchart-extension/package.json
@@ -21,10 +21,10 @@
     "start": "ts-node __tests__/runtime/browser/scripts/initVite.ts && vite serve __tests__/runtime/browser"
   },
   "dependencies": {
-    "@visactor/vrender-core": "~1.0.21",
-    "@visactor/vrender-kits": "~1.0.21",
-    "@visactor/vrender-components": "~1.0.21",
-    "@visactor/vrender-animate": "~1.0.21",
+    "@visactor/vrender-core": "~1.0.23",
+    "@visactor/vrender-kits": "~1.0.23",
+    "@visactor/vrender-components": "~1.0.23",
+    "@visactor/vrender-animate": "~1.0.23",
     "@visactor/vchart": "workspace:2.0.7",
     "@visactor/vutils": "~1.0.12",
     "@visactor/vdataset": "~1.0.12",
