+++ packages/vchart/package.json
@@ -122,10 +122,10 @@
     "@visactor/vdataset": "~1.0.12",
     "@visactor/vscale": "~1.0.12",
     "@visactor/vlayouts": "~1.0.12",
-    "@visactor/vrender-core": "~1.0.21",
-    "@visactor/vrender-kits": "~1.0.21",
-    "@visactor/vrender-components": "~1.0.21",
-    "@visactor/vrender-animate": "~1.0.21",
+    "@visactor/vrender-core": "~1.0.23",
+    "@visactor/vrender-kits": "~1.0.23",
+    "@visactor/vrender-components": "~1.0.23",
+    "@visactor/vrender-animate": "~1.0.23",
     "@visactor/vutils-extension": "workspace:2.0.7"
   },
   "publishConfig": {
