+++ tools/story-player/package.json
@@ -56,10 +56,10 @@
     "vite": "3.2.6"
   },
   "dependencies": {
-    "@visactor/vrender-core": "~1.0.21",
-    "@visactor/vrender-kits": "~1.0.21",
+    "@visactor/vrender-core": "~1.0.23",
+    "@visactor/vrender-kits": "~1.0.23",
     "@visactor/vchart": "workspace:2.0.7",
-    "@visactor/vrender": "~1.0.21",
+    "@visactor/vrender": "~1.0.23",
     "@visactor/vutils": "~1.0.12"
   }
 }
